rust: implement basic lifetimes for strings
Add an optional attribute to node.annotation. Targets that care
about it (such as rust in this diff and cpp and others in the future)
can specialize the types further.

+++ tests/expected/infer_ops.rs
@@ -20,7 +20,7 @@ pub fn foo() {
     let _c1: i32 = (a + b);
     let _c2: i32 = (a - b);
     let _c3: i32 = (a * b);
-    let _c4: f64 = ((a as f64) / (b as f64)) as f64;
+    let _c4: f64 = ((a as f64) / (b as f64));
     let _c5: i32 = -(a);
     let d: f64 = 2.0;
     let _e1: f64 = ((a as f64) + d);
